refactor: move User definition to pgdog-stats
