+++ pgdog-stats/src/lib.rs
@@ -6,9 +6,11 @@ pub mod resharding;
 pub mod schema;
 pub mod server;
 pub mod state;
+pub mod user;
 
 pub use memory::*;
 pub use pool::*;
 pub use replication::*;
 pub use resharding::*;
 pub use schema::*;
+pub use user::*;
