+++ pgdog-stats/src/user.rs
@@ -1,0 +1,16 @@
+use serde::{Deserialize, Serialize};
+
+/// Database/user pair that identifies a database cluster pool.
+#[derive(Debug, PartialEq, Hash, Eq, Clone, Default, Serialize, Deserialize)]
+pub struct User {
+    /// User name.
+    pub user: String,
+    /// Database name.
+    pub database: String,
+}
+
+impl std::fmt::Display for User {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}/{}", self.user, self.database)
+    }
+}
